+++ tests/metas.test.ts
@@ -1,0 +1,13 @@
+import { assertSiteSnapshot, build, getSite } from "./utils.ts";
+import metas from "../plugins/metas.ts";
+
+Deno.test("metas plugin", async (t) => {
+  const site = getSite({
+    src: "metas",
+  });
+
+  site.use(metas());
+
+  await build(site);
+  await assertSiteSnapshot(t, site);
+});